scraper/requirements.txt
scrapy
pymongo
pymongo[srv]
python-dotenv
pandas
dnspython
